Removed Json Converter Classes and refactored code into Json Classes and made the methods static, rendereing the converter classes obsolete. Need to refactor some of the logic in those methods before moving onto StoreConfigUpdate classes
