+++ Json Validator/JsonPopulator/JsonValidator/Json Classes/PurchaseScreenAppearance.cs
@@ -1,4 +1,6 @@
 ﻿using System.Collections.Generic;
+using System.Linq;
+using System.Windows.Forms;
 
 namespace JsonValidator
 {
@@ -19,6 +21,26 @@ namespace JsonValidator
 
             PopIds = addToMePopId;
             TitleLocKey = fandomName + "BoxTitle";
+        }
+        public static PurchaseScreenAppearance GeneratePurchaseScreenApeparance(Form1 form)
+        {
+            var flowBoxes = form.Controls.OfType<FlowLayoutPanel>().ToList();
+            var textBoxes = form.Controls.OfType<TextBox>().ToList();
+
+            var popLists = flowBoxes.Find(x => x.Name == "purchasePopsPanel").Controls.OfType<ComboBox>().ToList();
+            List<string> _popIds = new List<string>();
+
+            foreach (var x in popLists) //break me into a new class with store appearance
+            {
+                _popIds.Add(x.Text);
+            }
+            PurchaseScreenAppearance purchaseScreenAppearance = new PurchaseScreenAppearance()
+            {
+                TitleLocKey = textBoxes.Find(x => x.Name == "purTitleLocKey").Text,
+                PopIds = _popIds
+            };
+
+            return purchaseScreenAppearance;
         }
     }
