+++ Json Validator/JsonPopulator/JsonValidator/Json Classes/StoreButtonAppearance.cs
@@ -1,5 +1,8 @@
 ﻿using System.Collections.Generic;
 using System.Diagnostics;
+using System.Linq;
+using System.Windows.Forms;
+using System;
 
 namespace JsonValidator
 {
@@ -41,6 +44,32 @@ namespace JsonValidator
                     }
             }
             return orderedRarityList;
+        }
+        public static StoreButtonAppearance GenerateStoreBA(Form1 form)
+        {
+            var comboBoxes = form.Controls.OfType<ComboBox>().ToList();
+            var flowBoxes = form.Controls.OfType<FlowLayoutPanel>().ToList();
+
+            var popLists = flowBoxes.Find(x => x.Name == "storePopsPanel").Controls.OfType<ComboBox>().ToList();
+            List<string> _popIds = new List<string>();
+
+            foreach (var x in popLists) //break me into a new class with store appearance
+            {
+                _popIds.Add(x.Text);
+            }
+
+            StoreButtonAppearance storeButtonAppearance = new StoreButtonAppearance()
+            {
+                Style = comboBoxes.Find(x => x.Name == "styleCB").Text,
+                RibbonLocKey = comboBoxes.Find(x => x.Name == "ribbonLocKeyCB").Text,
+                TitleLocKey = comboBoxes.Find(x => x.Name == "titleLocCB").Text,
+                SubtitleLocKey = comboBoxes.Find(x => x.Name == "subLocCB").Text,
+                PopIds = _popIds,
+                Order = Convert.ToInt32(comboBoxes.Find(x => x.Name == "orderCB").Text),
+                Discount = Convert.ToInt32(comboBoxes.Find(x => x.Name == "discountCB").Text)
+            };
+
+            return storeButtonAppearance;
         }
     }
 }
