+++ Json Validator/JsonPopulator/JsonValidator/JsonConverters/PrizesConverter.cs
@@ -11,23 +11,23 @@ namespace JsonValidator
 {
     public class PrizesConverter
     {
-        public List<Prize> GeneratePrizeList(List<FlowLayoutPanel> flowList)
-        {
-            var popLists = flowList.Find(x => x.Name == "prizePanel").Controls.OfType<GroupBox>().ToList();
-            List<Prize> prizePopList = new List<Prize>();
+        //public List<Prize> GeneratePrizeList(List<FlowLayoutPanel> flowList)
+        //{
+        //    var popLists = flowList.Find(x => x.Name == "prizePanel").Controls.OfType<GroupBox>().ToList();
+        //    List<Prize> prizePopList = new List<Prize>();
 
-            foreach (var x in popLists)
-            {
-                Prize prize = new Prize();
-                var comboB = x.Controls.OfType<ComboBox>().ToList();
-                prize.RewardType = comboB[0].Text;
-                prize.RewardID = comboB[1].Text;
-                prize.Amount = Convert.ToInt32(comboB[2].Text);
-                prize.Instances = Convert.ToInt32(comboB[3].Text);
-                prizePopList.Add(prize);
-            }
+        //    foreach (var x in popLists)
+        //    {
+        //        Prize prize = new Prize();
+        //        var comboB = x.Controls.OfType<ComboBox>().ToList();
+        //        prize.RewardType = comboB[0].Text;
+        //        prize.RewardID = comboB[1].Text;
+        //        prize.Amount = Convert.ToInt32(comboB[2].Text);
+        //        prize.Instances = Convert.ToInt32(comboB[3].Text);
+        //        prizePopList.Add(prize);
+        //    }
 
-            return prizePopList;
-        }
+        //    return prizePopList;
+        //}
     }
 }
